Moved takeScreenshot method to AbstractWebDriverCore
